fix bad routing inheritance, rename `RoutingModelBase` to `EtheruemRoutingModel`
